useValueIterator subscribes only on mount
