+++ src/create-element/parse-component.ts
@@ -54,7 +54,6 @@ function parseComponent({
             componentAPI.onUnmount(mountCbResult);
           }
         });
-        componentMountCbs = [];
       },
       _callUnmountHandlers: () => {
         // this should trigger recursive checks, whether it is a VelesNode or VelesComponent
@@ -65,7 +64,6 @@ function parseComponent({
 
         // we execute own unmount callbacks after children, so the order is reversed
         componentUnmountCbs.forEach((cb) => cb());
-        componentUnmountCbs = [];
       },
     },
   };
